Use a dedicated quiet terminal launcher
